Renamed some stuff, improved the example program

+++ Collision2D/BoundaryElement.cpp
@@ -34,7 +34,7 @@ pointSeparationData_type BoundaryElement::calculateMinPointSeparation(sf::Vector
     return separationData;
 }
 
-int BoundaryElement::getNormalCount() {
+size_t BoundaryElement::getNormalCount() {
     // There is only one normal vector
     return 1;
 }
